fix(TCPHeader): correct the way headerLen calculated

+++ header/header.go
@@ -22,7 +22,7 @@ type IPHeader interface {
 type ProtocolHeader interface {
 	String() string
 
-	HeaderLen() uint8
+	HeaderLen() int
 	Checksum() uint16
 	SrcPort() (uint16, error)
 	DstPort() (uint16, error)
